ai_model/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
async-timeout==5.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
distro==1.9.0
docx2txt==0.9
EbookLib==0.19
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fsspec==2025.5.1
groq==0.28.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.65
langchain-groq==0.3.2
langsmith==0.3.45
lxml==5.4.0
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.2.6
openai==1.88.0
orjson==3.10.18
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.1.0
python-pptx==1.0.2
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
tenacity==9.1.2
textract==1.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
xlrd==2.0.2
xlsxwriter==3.2.5
zstandard==0.23.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
langchain-groq>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
# Web Speech API is browser-native - no Python package required
# speech.py uses window.SpeechRecognition (frontend JS only)
